# [enhancement](Nereids): optimize deepEquals