## Proposed changes

A big query

First Time: 1.953s
Multiple Time: 0.544s

First Time: 1.738s
Multiple Time: 0.475s

<!--Describe your changes.-->

## Further comments

If this is a relatively large or complex change, kick off the discussion at [dev@doris.apache.org](mailto:dev@doris.apache.org) by explaining why you chose the solution you did and what alternatives you considered, etc...